requirements.txt
feedparser>=6.0.11
requests>=2.31.0
python-dotenv>=1.0.1
google-genai>=1.16.1
